fix: project section: dev
